chore: Update release workflow to build and upload artifacts
The release workflow has been updated to build and upload artifacts. Previously, it only triggered on published releases, but now it will trigger on any created release. The workflow now includes a step to compile and release the code using the rust-build action. Additionally, an upload artifact step has been added to upload the built binary and checksum files.